lpm: Check secondary core pm state before lpm entry
During suspend secondary core is turned off in parallel to the lpm
sequence. Because of which there is a race condition where lpm
sequence can start before the secondary A53 core is turned off. Add
a check before starting the lpm sequence see if the secondary core
is turned off or not.

Signed-off-by: Sebin Francis <sebin.francis@ti.com>