FROMLIST: refactor(ti): Introduce a top level TI SCI Transport layer
There maybe multiple ways to transport TI SCI messages, hence move all the
k3_sec_proxy_XYZ calls into a common TI SCI Transport header.
Also add support for consuming boot notification as it's a TI SCI message
defined in the TISCI General Message API Documentation [1].
This message is special in the sense that it notifies if system-firmware is
ready to receive TISCI messages. Hence, it is required to be put in the
bl31_setup as well.

With this, we can better abstract the lower level transport layers from the
callers of ti_sci_transport_XYZ.

While at it, rename all the instances in ti_sci to use the new APIs.

References:
[1] https://software-dl.ti.com/tisci/esd/09_02_07/2_tisci_msgs/general/core.html#tisci-msg-boot-notification

Link:
https://review.trustedfirmware.org/c/TF-A/trusted-firmware-a/+/33998/4
Change-Id: I05673f379b1398c0b6a2bd9e1b5392165d12f151
Signed-off-by: Dhruva Gole <d-gole@ti.com>