FROMLIST: feat(ti): Add support for scmi power and clock mgmt
Add scmi_pd,clk files which are responsible for the scmi protocol handling
for the clock/ power operation requests coming from SCMI client.
Also, Add the pd and clk arrays for AM62L into it's own folder so
that the scmi drivers can be generic enough and pick the soc specific
arrays from the soc folder data files.

Also enable SCMI PD and clock by building those files
and with the new files we require to increase the BL31_SIZE

Link:
https://review.trustedfirmware.org/c/TF-A/trusted-firmware-a/+/34834/1
Change-Id: I1e07b1c8423d3f273552caeda51f3225b75e1419
Co-developed-by: Kamlesh Gurudasani <kamlesh@ti.com>
Signed-off-by: Kamlesh Gurudasani <kamlesh@ti.com>
Signed-off-by: Dhruva Gole <d-gole@ti.com>