feat(k3): Encrypt TFA at suspend
The TIFS is able to encrypt BL31 to a DDR location at suspend.
Then, it will be resored at resume by a decrypt message sent by the SPL
on k3 plartforms.

Here, the TIFS doesn't support ti_sci_lpm_get_next_sys_mode(), so we
force the mode to SOC_OFF, which is the only one accepted by TIFS.

Also, when sending ti_sci_query_fw_caps(), the TIFS only respond with
MSG_FLAG_CAPS_LPM_DM_MANAGED, so we need to invert the test, but it
would be better if the TIFS could send a SOC_OFF CAPS maybe.

Signed-off-by: Richard Genoud <richard.genoud@bootlin.com>
Signed-off-by: Abhash Kumar Jha <a-kumar2@ti.com>