feat(ti): lpm: Add LPM entry and exit sequence
Add PSCI handlers for LPM entry and exit. In the handlers call the low
level LPM sequence handlers.
LPM STUB will be executing the low level LPM sequence. it will keep the
DDR in self refresh. Because of which it cannot run from DDR. Copy the
stub from DDR to wkup sram. Add wkup sram in the mmu map table. Increase
the BL31 size to accommodate the LPM stub.

Change-Id: Ia10b04787c842efe2eedccb653d928c97b15b592
Signed-off-by: Sebin Francis <sebin.francis@ti.com>