PENDING: feat(ti): lpm: Add lpm related init sequences
RTC used as a wake up source. Initialize the RTC during boot.

LPM STUB will be executing the low level LPM sequence. it will keep the
DDR in self refresh. Because of which it cannot run from DDR. Copy the
stub from DDR to wkup sram. Add wkup sram in the mmu map table. Increase
the BL31 size to accommodate the LPM stub.

Change-Id: Idfaa681394371afb489a96b083758d12a5910616
Signed-off-by: Sebin Francis <sebin.francis@ti.com>