PENDING: fix(ti): Flush OTP memory blob before handoff
The OTP blob is ultimately consumed by the secure co-processor in the
system and thus needs to be flushed to the point of cache coherency.
These blobs are at most a couple of KBs in size, but flush the entire
PAGE to be on the safer side.

Without this change, the responsibility for flushing the cache for the
blob being passed lies with the caller (like U-Boot), which isn't the
right place since U-Boot has no knowledge of the backend implementation
of the SVC call.

Fixes: 7a76ef2f283d ("PENDING: feat(ti): add keywriter lite SMC handler")
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>