feat(ti): tisci: Add support for get_suspend_controller message
During LPM sequence TFA needs to know which cpu is the suspend controller.
Add support to send get_suspend_controller message to TIFS

If the A72 cpu initiates the suspend, then we add a `TI_SCI_FLAG_REQ_FWD2DM`
flag to the enter sleep message. Based on if this flag is set or not
TIFS can forward the enter_sleep message to the DM firmware.

Signed-off-by: Abhash Kumar Jha <a-kumar2@ti.com>