PENDING: feat(ti): Simplify firewall configuration on AM62L
Commit 5939ceaeb ("PENDING: feat(ti): Re-configure firewalls for AM62L")
overcomes the ROM configurations of firewalls by adding a foreground
configuration for each fwl ID. This also needed to check the device type
and add the new configs for only HS-SE devices.

While this works, a much simpler approach is to just disable the
offending background firewall configurations. We do not need to manually
add specific memory ranges to a foreground configuration to overwrite
the background configurations. This is already being done for some other
SoCs in U-Boot. These firewalls are enabled by ROM for boot phase, and
are now not needed after boot is done. As an added benefit, HS-FS
devices also boot perfectly fine with this change. So this commit also
removes the device type check code added before. Tested on both HS-FS
and HS-SE boards.

Signed-off-by: T Pratham <t-pratham@ti.com>